Implement write(name, value) for ConsolePaper
As there are use cases where a form is pre-filled with values
the paper needs to be able to accept those values before asking
the user for it's input. Furthermore it should reuse those values
when the user does not want to change them.
This is now implemented by adding the message 'askForInput'.

This fixes #33